+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - EEPROM dump support, fix #12
 - Refactor all subcommands, using clap v4
 - Probe support, multiple chips can be selected by an index
+- Progressbar for flash and verify commands
+
 ### Changed
 - Disable debug log by default